pyproject.toml
# ==============================================================
# Stabilo - MIT License
# ==============================================================
# Overview:
# This pyproject.toml file configures the build system and 
# project metadata for the Stabilo package. 
#
# Installation:
# To install the Stabilo package, use the following command:
#   pip install stabilo
#
# For development purposes, enabling real-time updates without 
# reinstallation, clone the repository and install the package
# in editable mode with the following command:
#   pip install -e .[dev]
# or for zsh shell:
#   pip install -e '.[dev]'
#
# ==============================================================

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stabilo"
dynamic = ["version"]
description = "Stabilizes video or extracted trajectories with respect to a selected reference frame in the video, with optional user-provided masks."
authors = [
    { name = "Robert Fonod", email = "robert.fonod@ieee.org" },
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = ["stabilo", "video-stabilization", "object-stabilization", "mask", "reference-frame", "computer-vision"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "numpy>=1.26.4,<2.0",
    "opencv-python>=4.6.0",
    "pyyaml>=5.3.1",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
extras = [
    "matplotlib>=3.5.0"
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.11.0",
    "matplotlib>=3.5.0"
]

[tool.setuptools] 
packages = { find = { where = ["."], include = ["stabilo", "stabilo.*"] } }
package-data = { "stabilo" = ["**/*.yaml", "**/*.txt"] }

[tool.setuptools.dynamic]
version = { attr = "stabilo.__version__" }

[project.urls]
Homepage = "https://github.com/rfonod/stabilo/"
Repository = "https://github.com/rfonod/stabilo/"
Changelog = "https://github.com/rfonod/stabilo/releases"
Issues = "https://github.com/rfonod/stabilo/issues/"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ['E', 'F', 'W', 'A', "B", 'PLC', 'PLE', 'PLW', 'I']

[tool.ruff.format]
docstring-code-format = true
quote-style = "preserve"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = [
    "tests",
]